Restructure Module
* Removes `pkg`
* Moves all code to `internal` except for the intentional public API
* Moves main.go for CLI tool to `cmd`
