+++ internal/gguf/create.go
@@ -8,8 +8,8 @@ import (
 	v1 "github.com/google/go-containerregistry/pkg/v1"
 	gguf_parser "github.com/gpustack/gguf-parser-go"
 
-	"github.com/docker/model-distribution/pkg/partial"
-	"github.com/docker/model-distribution/pkg/types"
+	"github.com/docker/model-distribution/internal/partial"
+	"github.com/docker/model-distribution/types"
 )
 
 func NewModel(path string) (*Model, error) {
