+++ internal/mutate/mutate.go
@@ -4,7 +4,7 @@ import (
 	v1 "github.com/google/go-containerregistry/pkg/v1"
 	ggcr "github.com/google/go-containerregistry/pkg/v1/types"
 
-	"github.com/docker/model-distribution/pkg/types"
+	"github.com/docker/model-distribution/types"
 )
 
 func AppendLayers(mdl types.ModelArtifact, layers ...v1.Layer) types.ModelArtifact {
