+++ internal/store/errors.go
