+++ internal/store/testdata/dummy.gguf
