+++ internal/store/testdata/license.txt
