+++ types/model.go
@@ -1,6 +1,8 @@
 package types
 
-import v1 "github.com/google/go-containerregistry/pkg/v1"
+import (
+	v1 "github.com/google/go-containerregistry/pkg/v1"
+)
 
 type Model interface {
 	ID() (string, error)
